requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-allauth>=65.0
django-axes>=6.0
django-redis>=5.0
psycopg2-binary>=2.9
django-htmx>=1.17
whitenoise>=6.5
django-environ>=0.11
requests>=2.31
PyJWT>=2.8
cryptography>=42

# Dev / QA
pytest>=8.0
pytest-django>=4.8
factory_boy>=3.3
coverage>=7.6
black>=24.0
isort>=5.13
flake8>=7.0
mypy>=1.10
types-requests>=2.32
mysqlclient>=2.2
